add anki-helper and improve screenshot script

+++ pkgs/screenshot.nix
@@ -40,7 +40,8 @@ writeShellScriptBin "screenshot" ''
   "current window")
     [[ $rofi_used == 1 ]] && sleep 0.2
     output=$(mmsg -g -o | grep "selmon 1" | cut -d' ' -f1)
-    geometry=$(mmsg -x -o "$output" | awk '/^x /{x=$2} /^y /{y=$2} /^width /{w=$2} /^height /{h=$2} END{print x","y" "w"x"h}')
+    geometry=$(mmsg -x | grep "$output" | awk '{vals[$2]=$3} END{print vals["x"]","vals["y"]" "vals["width"]"x"vals["height"]}')
+    echo $geometry
     grim -g "$geometry" /tmp/grim_screenshot.png
     ;;
   *) exit ;;
@@ -48,27 +49,33 @@ writeShellScriptBin "screenshot" ''
 
   file -f /tmp/grim_screenshot.png >/dev/null 2>&1 && cat /tmp/grim_screenshot.png | wl-copy
 
-  while true; do
-    name=$(rofi -dmenu -p "Filename" -lines 0 -width 30)
 
-    if [ "$name" == "" ]; then
-      cancel=$(echo -en "No\nYes" | rofi -dmenu -p "Cancel?" -lines 2 -width 30)
-      if [ "$cancel" != "No" ]; then
-        rm /tmp/grim_screenshot.png
-        exit
+  if [[ -n "$2" ]]; then
+    out_path="$2"
+  else
+    while true; do
+      name=$(rofi -dmenu -p "Filename" -lines 0 -width 30)
+
+      if [ "$name" == "" ]; then
+        cancel=$(echo -en "No\nYes" | rofi -dmenu -p "Cancel?" -lines 2 -width 30)
+        if [ "$cancel" != "No" ]; then
+          rm /tmp/grim_screenshot.png
+          exit
+        fi
       fi
-    fi
 
-    if ! test -f ~/Pictures/"$name".png; then
-      break
-    fi
+      if ! test -f ~/Pictures/"$name".png; then
+        break
+      fi
 
-    overwrite=$(echo -en "No\nYes" | rofi -dmenu -p "Do you want to overwrite $name.png?" -lines 2 -width 30)
+      overwrite=$(echo -en "No\nYes" | rofi -dmenu -p "Do you want to overwrite $name.png?" -lines 2 -width 30)
 
-    if [ "$overwrite" == "Yes" ]; then
-      break
-    fi
-  done
+      if [ "$overwrite" == "Yes" ]; then
+        break
+      fi
+    done
+    out_path = ~/Pictures/"$name".png
+  fi
 
-  mv /tmp/grim_screenshot.png ~/Pictures/"$name".png
+  mv /tmp/grim_screenshot.png $out_path
 ''
